backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Jobs
celery==5.3.4
redis==5.0.1
flower==2.0.1

# AI & LLM
openai==1.10.0
langchain==0.1.5
chromadb==0.4.22
sentence-transformers==2.3.1

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.17.2
requests==2.31.0
lxml==5.1.0

# Payment Processing
stripe==8.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
email-validator==2.1.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
